Moved pre-made propulsion layouts from examples folder to propulsion
