+++ openconcept/examples/Caravan.py
@@ -6,7 +6,7 @@ import openconcept.api as oc
 
 # imports for the airplane model itself
 from openconcept.examples.aircraft_data.caravan import data as acdata
-from openconcept.examples.propulsion_layouts.simple_turboprop import TurbopropPropulsionSystem
+from openconcept.propulsion import TurbopropPropulsionSystem
 from openconcept.examples.methods.weights_turboprop import SingleTurboPropEmptyWeight
 from openconcept.examples.methods.costs_commuter import OperatingCost
