+++ openconcept/examples/ElectricSinglewithThermal.py
@@ -11,7 +11,7 @@ from openconcept.utilities.math import AddSubtractComp
 from openconcept.utilities.math.integrals import Integrator
 from openconcept.utilities.dvlabel import DVLabel
 from openconcept.examples.methods.weights_turboprop import SingleTurboPropEmptyWeight
-from openconcept.examples.propulsion_layouts.simple_all_electric import AllElectricSinglePropulsionSystemWithThermal_Incompressible, AllElectricSinglePropulsionSystemWithThermal_Compressible
+from openconcept.propulsion import AllElectricSinglePropulsionSystemWithThermal_Incompressible, AllElectricSinglePropulsionSystemWithThermal_Compressible
 from openconcept.examples.methods.costs_commuter import OperatingCost
 from openconcept.utilities.dict_indepvarcomp import DictIndepVarComp
 from openconcept.examples.aircraft_data.TBM850 import data as acdata
