+++ openconcept/examples/HybridTwin_active_thermal.py
@@ -14,7 +14,7 @@ from openconcept.utilities.math.max_min_comp import MaxComp
 from openconcept.utilities.math.integrals import Integrator
 from openconcept.utilities.dvlabel import DVLabel
 from openconcept.examples.methods.weights_twin_hybrid import TwinSeriesHybridEmptyWeight
-from openconcept.examples.propulsion_layouts.thermal_series_hybrid import TwinSeriesHybridElectricPropulsionRefrigerated
+from openconcept.propulsion import TwinSeriesHybridElectricThermalPropulsionRefrigerated
 from openconcept.examples.methods.costs_commuter import OperatingCost
 from openconcept.utilities.dict_indepvarcomp import DictIndepVarComp
 from openconcept.examples.aircraft_data.KingAirC90GT import data as acdata
@@ -24,7 +24,7 @@ from openconcept.utilities.visualization import plot_trajectory
 
 """
 WARNING: This example has known convergence problems because of the chiller in the
-         propulsion layout (TwinSeriesHybridElectricPropulsionRefrigerated)
+         propulsion layout (TwinSeriesHybridElectricThermalPropulsionRefrigerated)
 
 Eytan Adler, 27/10/2021
 """
@@ -73,7 +73,7 @@ class SeriesHybridTwinModel(Group):
                                       "ac|weights*", 'duration']
 
         self.add_subsystem('propmodel',
-                           TwinSeriesHybridElectricPropulsionRefrigerated(num_nodes=nn),
+                           TwinSeriesHybridElectricThermalPropulsionRefrigerated(num_nodes=nn),
                            promotes_inputs=propulsion_promotes_inputs,
                            promotes_outputs=propulsion_promotes_outputs)
         self.connect('proprpm', ['propmodel.prop1.rpm', 'propmodel.prop2.rpm'])
