+++ openconcept/examples/TBM850.py
@@ -4,7 +4,7 @@ import openmdao.api as om
 import matplotlib.pyplot as plt
 
 # OpenConcept imports for the airplane model
-from openconcept.examples.propulsion_layouts.simple_turboprop import TurbopropPropulsionSystem
+from openconcept.propulsion import TurbopropPropulsionSystem
 from openconcept.aerodynamics import PolarDrag
 from openconcept.examples.methods.weights_turboprop import SingleTurboPropEmptyWeight
 from openconcept.utilities.math import AddSubtractComp
