+++ openconcept/propulsion/systems/simple_all_electric.py
