+++ openconcept/propulsion/systems/simple_turboprop.py
